fix(test): rename example plugin
