+++ test/fixtures/plugin.textlintrc
@@ -1,6 +1,6 @@
 {
   "plugins": [
-    "example-plugin",
+    "example",
     "configurable-plugin"
   ]
 }
