+++ test/fixtures/plugins/textlint-plugin-example/index.js
