+++ test/fixtures/plugins/textlint-plugin-example/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "example-plugin",
+  "name": "textlint-plugin-example",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
