+++ test/fixtures/plugins/textlint-plugin-example/rules/example-rule.js
